GraphQL endpoint changed to require authorization for modification operations

+++ mapr-rest/src/main/java/com/mapr/music/api/graphql/GraphQLEndpoint.java
@@ -5,16 +5,20 @@ import com.mapr.music.dto.AlbumDto;
 import com.mapr.music.dto.ArtistDto;
 import com.mapr.music.service.AlbumService;
 import com.mapr.music.service.ArtistService;
+import com.mapr.music.service.UserService;
+import graphql.GraphQLException;
 import graphql.schema.GraphQLSchema;
 import graphql.schema.idl.RuntimeWiring;
 import graphql.schema.idl.SchemaGenerator;
 import graphql.schema.idl.SchemaParser;
 import graphql.schema.idl.TypeDefinitionRegistry;
 import graphql.servlet.SimpleGraphQLServlet;
+import org.jboss.resteasy.spi.ResteasyProviderFactory;
 
 import javax.inject.Inject;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.security.Principal;
 import java.util.stream.Collectors;
 
 public class GraphQLEndpoint extends SimpleGraphQLServlet {
@@ -24,22 +28,29 @@ public class GraphQLEndpoint extends SimpleGraphQLServlet {
   private static final ObjectMapper MAPPER = new ObjectMapper();
 
   @Inject
-  public GraphQLEndpoint(AlbumService albumService, ArtistService artistService) {
-    super(createGraphQLSchema(albumService, artistService));
+  public GraphQLEndpoint(AlbumService albumService, ArtistService artistService, UserService userService) {
+    super(createGraphQLSchema(albumService, artistService, userService));
   }
 
-  private static GraphQLSchema createGraphQLSchema(AlbumService albumService, ArtistService artistService) {
+  private static GraphQLSchema createGraphQLSchema(AlbumService albumService, ArtistService artistService, UserService userService) {
 
     SchemaParser schemaParser = new SchemaParser();
     InputStream schemaFile = GraphQLEndpoint.class.getClassLoader().getResourceAsStream(SCHEMA_FILENAME);
     TypeDefinitionRegistry typeRegistry = schemaParser.parse(new InputStreamReader(schemaFile));
 
-    return new SchemaGenerator().makeExecutableSchema(typeRegistry, buildRuntimeWiring(albumService, artistService));
+    return new SchemaGenerator().makeExecutableSchema(typeRegistry, buildRuntimeWiring(albumService, artistService, userService));
   }
 
-  private static RuntimeWiring buildRuntimeWiring(AlbumService albumService, ArtistService artistService) {
+  private static RuntimeWiring buildRuntimeWiring(AlbumService albumService, ArtistService artistService, UserService userService) {
     return RuntimeWiring.newRuntimeWiring()
       .type("Query", typeWiring -> typeWiring
+        .dataFetcher("currentUser", (env) -> {
+          Principal principal = ResteasyProviderFactory.getContextData(Principal.class);
+          if (principal == null) {
+            throw new GraphQLException("Unauthorized");
+          }
+          return userService.getUserByPrincipal(principal);
+        })
         .dataFetcher("album", (env) -> albumService.getAlbumById(env.getArgument("id")))
         .dataFetcher("albums", (env) -> {
           Integer offset = env.getArgument("offset");
@@ -54,19 +65,29 @@ public class GraphQLEndpoint extends SimpleGraphQLServlet {
         })
       )
       .type("Mutation", typeWiring -> typeWiring
-        .dataFetcher("createAlbum",
-          (env) -> albumService.createAlbum(MAPPER.convertValue(env.getArgument("album"), AlbumDto.class)))
-        .dataFetcher("updateAlbum",
-          (env) -> albumService.updateAlbum(MAPPER.convertValue(env.getArgument("album"), AlbumDto.class)))
+        .dataFetcher("createAlbum", (env) -> {
+          assertAuthorized();
+          return albumService.createAlbum(MAPPER.convertValue(env.getArgument("album"), AlbumDto.class));
+        })
+        .dataFetcher("updateAlbum", (env) -> {
+          assertAuthorized();
+          return albumService.updateAlbum(MAPPER.convertValue(env.getArgument("album"), AlbumDto.class));
+        })
         .dataFetcher("deleteAlbum", (env) -> {
+          assertAuthorized();
           albumService.deleteAlbumById(env.getArgument("id"));
           return true;
         })
-        .dataFetcher("createArtist",
-          (env) -> artistService.createArtist(MAPPER.convertValue(env.getArgument("artist"), ArtistDto.class)))
-        .dataFetcher("updateArtist",
-          (env) -> artistService.updateArtist(MAPPER.convertValue(env.getArgument("artist"), ArtistDto.class)))
+        .dataFetcher("createArtist", (env) -> {
+          assertAuthorized();
+          return artistService.createArtist(MAPPER.convertValue(env.getArgument("artist"), ArtistDto.class));
+        })
+        .dataFetcher("updateArtist", (env) -> {
+          assertAuthorized();
+          return artistService.updateArtist(MAPPER.convertValue(env.getArgument("artist"), ArtistDto.class));
+        })
         .dataFetcher("deleteArtist", (env) -> {
+          assertAuthorized();
           artistService.deleteArtistById(env.getArgument("id"));
           return true;
         })
@@ -94,5 +115,12 @@ public class GraphQLEndpoint extends SimpleGraphQLServlet {
         })
       )
       .build();
+  }
+
+  private static void assertAuthorized() {
+    Principal principal = ResteasyProviderFactory.getContextData(Principal.class);
+    if (principal == null) {
+      throw new GraphQLException("Unauthorized");
+    }
   }
 }
